+++ src/SharpX/Extensions/ObjectExtensions.cs
@@ -1,0 +1,7 @@
+namespace SharpX.Extensions;
+
+public static class ObjectExtensions
+{
+    /// <summary>Discards a value and return <c>Unit</c>.</summary>
+    public static Unit ToUnit<T>(this T value) => Unit.Default;
+}